// web/src/api/models/home-docker-service.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { HomeDockerStats } from './home-docker-stats';

/**
 * 
 * @export
 * @interface HomeDockerService
 */
export interface HomeDockerService {
    /**
     * 
     * @type {HomeDockerStats}
     * @memberof HomeDockerService
     */
    'log_stream': HomeDockerStats;
    /**
     * 
     * @type {HomeDockerStats}
     * @memberof HomeDockerService
     */
    'ocserv': HomeDockerStats;
    /**
     * 
     * @type {HomeDockerStats}
     * @memberof HomeDockerService
     */
    'postgres': HomeDockerStats;
    /**
     * 
     * @type {HomeDockerStats}
     * @memberof HomeDockerService
     */
    'user_expiry': HomeDockerStats;
    /**
     * 
     * @type {HomeDockerStats}
     * @memberof HomeDockerService
     */
    'web': HomeDockerStats;
}