// web/src/api/models/home-get-home-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ModelsOnlineUserSession } from './models-online-user-session';

/**
 * 
 * @export
 * @interface HomeGetHomeUser
 */
export interface HomeGetHomeUser {
    /**
     * 
     * @type {Array<ModelsOnlineUserSession>}
     * @memberof HomeGetHomeUser
     */
    'online_users_session'?: Array<ModelsOnlineUserSession>;
    /**
     * 
     * @type {number}
     * @memberof HomeGetHomeUser
     */
    'total'?: number;
}