// web/src/api/models/models-daily-traffic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ModelsDailyTraffic
 */
export interface ModelsDailyTraffic {
    /**
     * Format: YYYY-MM-DD
     * @type {string}
     * @memberof ModelsDailyTraffic
     */
    date?: string;
    /**
     * in GiB
     * @type {number}
     * @memberof ModelsDailyTraffic
     */
    rx?: number;
    /**
     * in GiB
     * @type {number}
     * @memberof ModelsDailyTraffic
     */
    tx?: number;
}