// web/src/api/models/models-ipban-points.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ModelsIPBanPoints
 */
export interface ModelsIPBanPoints {
    /**
     *
     * @type {string}
     * @memberof ModelsIPBanPoints
     */
    IP?: string;
    /**
     *
     * @type {number}
     * @memberof ModelsIPBanPoints
     */
    Score?: number;
    /**
     *
     * @type {string}
     * @memberof ModelsIPBanPoints
     */
    Since?: string;
    /**
     *
     * @type {string}
     * @memberof ModelsIPBanPoints
     */
    _Since?: string;
}