// web/src/api/models/models-ocserv-group.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { ModelsOcservGroupConfig } from './models-ocserv-group-config';

/**
 *
 * @export
 * @interface ModelsOcservGroup
 */
export interface ModelsOcservGroup {
    /**
     *
     * @type {ModelsOcservGroupConfig}
     * @memberof ModelsOcservGroup
     */
    config?: ModelsOcservGroupConfig;
    /**
     *
     * @type {number}
     * @memberof ModelsOcservGroup
     */
    id?: number;
    /**
     *
     * @type {string}
     * @memberof ModelsOcservGroup
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ModelsOcservGroup
     */
    owner: string;
}