// web/src/api/models/models-online-user-session.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ocserv User management Example Api
 * This is a sample Ocserv User management Api server.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ModelsOnlineUserSession
 */
export interface ModelsOnlineUserSession {
    /**
     *
     * @type {string}
     * @memberof ModelsOnlineUserSession
     */
    'Average RX'?: string;
    /**
     *
     * @type {string}
     * @memberof ModelsOnlineUserSession
     */
    'Average TX'?: string;
    /**
     *
     * @type {string}
     * @memberof ModelsOnlineUserSession
     */
    Groupname?: string;
    /**
     *
     * @type {string}
     * @memberof ModelsOnlineUserSession
     */
    Username?: string;
    /**
     *
     * @type {string}
     * @memberof ModelsOnlineUserSession
     */
    '_Connected at'?: string;
}